Rename: /redis/excption to /redis/exception
