+++ src/test/java/com/example/sendowl/redis/RedisRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.example.sendowl.redis;
 
 import com.example.sendowl.redis.entity.RedisBoard;
-import com.example.sendowl.redis.excption.RedisBoardNotFoundException;
+import com.example.sendowl.redis.exception.RedisBoardNotFoundException;
 import com.example.sendowl.redis.repository.RedisBoardRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
